requirements.txt
google-genai
python-dotenv
PyMuPDF
reportlab
python-docx
sqlalchemy
pillow
pytest
pyyaml
flask
openai
anthropic[vertex]
gunicorn

# Development tools (not required for production)
# ruff - linter and formatter: pip install ruff
